akili_doctor/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-admin==2.0.2
django-axes==7.0.0
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
excel-base==1.0.4
isoweek==1.3.3
pillow==11.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
screen==1.0.1
six==1.16.0
sqlparse==0.5.2
TimeConvert==3.0.13
tzlocal==5.2
xlwt==1.3.0
